allow autogold.Want calls anywhere in the test file
Prior to this change `autogold.Want` calls could only be located
under a `TestFoo` function in the `_test.go` file. This meant the
name passed to `autogold.Want(...)` only had to be unique within
the scope of the test function (rather than in the entire file)
but lead to confusion like in #7

This changes the behavior such that `autogold.Want` is allowed _anywhere_
in the test file, but the name must be unique within the scope of the
entire file rather than unique just in the test functions scope.

However, the real plan here is to remove the need for passing a name
parameter to autogold.Want(...) all together. This is part 1 of that change:
https://github.com/hexops/autogold/issues/11

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>