Fix -update with Colliding Subtest Names
Golden files are moved to a temp dir when -update is enabled, which
allows clean-up of golden files which are not in use. The destination
path in the temp dir did not consider the adjusted output directory
for subtests, which meant that subtests with the same name would be
copied to the same path in the temp dir, overwriting each other.
This conflict leads to a number of issues, which results in tests
failing despite producing the correct output.

By adjusting the destination path inside the temp dir to match the
relative dir used for the source (of the golden files), no conflict
occurs when moving them to/from the temp dir.