handle multiple (possibly parallel) inline replacements
Tests can run in parallel, and desire rewriting inline `autogold.Expect` calls at the same time.
As a result, we need to handle locking the file (which we now do with a PID lock) and keeping
track of line changes within the file as our own changes to the file may invalidate the line number
we look for to replace the `autogold.Expect` call based on call stack information.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>