allow usage together with other golden packages
Other packages can also define an -update flag. In general we can't be
compatible with every possible conflicting flag, but for this specific
flag we assume that it's a boolean flag with the same meaning as ours.

My motivation for this change is usage of autogold along with
gotest.tools/v3/assert, which among other things, has its own golden
tests functionality.