fast path for equal in Expect.Equal
This fast path is avoiding the need to call stringify if we know for
certain the values are equal. If we aren't certain we fallback to the
normal paths of comparing the stringified values.

Context: in the Sourcegraph repository a unit test goes from 4s to 0.1s
after this change.

Additionally this change includes a run of "go mod tidy" which was
needed for me to run "go test".

Test Plan: all tests pass in this repository. Additionally tests
continue to pass in the Sourcegraph repository after this change.